install.sh

+++ install.sh
@@ -8,8 +8,20 @@ plain='\033[0m'
 
 cur_dir=$(pwd)
 
+# Logging functions
+LOGE() {
+    echo -e "${red}[ERROR] $* ${plain}"
+}
+
+LOGI() {
+    echo -e "${green}[INFO] $* ${plain}"
+}
+
+log_file="/var/log/x-sl-install.log"
+touch $log_file
+
 # check root
-[[ $EUID -ne 0 ]] && echo -e "${red}Fatal error: ${plain} Please run this script with root privilege \n " && exit 1
+[[ $EUID -ne 0 ]] && LOGE "Fatal error: Please run this script with root privilege \n" && exit 1
 
 # Check OS and set release variable
 if [[ -f /etc/os-release ]]; then
@@ -19,10 +31,10 @@ elif [[ -f /usr/lib/os-release ]]; then
     source /usr/lib/os-release
     release=$ID
 else
-    echo "Failed to check the system OS, please contact the author!" >&2
+    LOGE "Failed to check the system OS, please contact the author!"
     exit 1
 fi
-echo "The OS release is: $release"
+LOGI "The OS release is: $release"
 
 arch() {
     case "$(uname -m)" in
@@ -33,65 +45,74 @@ arch() {
     armv6* | armv6) echo 'armv6' ;;
     armv5* | armv5) echo 'armv5' ;;
     s390x) echo 's390x' ;;
-    *) echo -e "${green}Unsupported CPU architecture! ${plain}" && rm -f install.sh && exit 1 ;;
+    *) LOGE "Unsupported CPU architecture: $(uname -m)" && exit 1 ;;
     esac
 }
 
-echo "arch: $(arch)"
+LOGI "arch: $(arch)"
 
 os_version=""
 os_version=$(grep "^VERSION_ID" /etc/os-release | cut -d '=' -f2 | tr -d '"' | tr -d '.')
 
 if [[ "${release}" == "arch" ]]; then
-    echo "Your OS is Arch Linux"
+    LOGI "Your OS is Arch Linux"
 elif [[ "${release}" == "parch" ]]; then
-    echo "Your OS is Parch Linux"
+    LOGI "Your OS is Parch Linux"
 elif [[ "${release}" == "manjaro" ]]; then
-    echo "Your OS is Manjaro"
+    LOGI "Your OS is Manjaro"
 elif [[ "${release}" == "armbian" ]]; then
-    echo "Your OS is Armbian"
+    LOGI "Your OS is Armbian"
 elif [[ "${release}" == "alpine" ]]; then
-    echo "Your OS is Alpine Linux"
+    LOGI "Your OS is Alpine Linux"
 elif [[ "${release}" == "opensuse-tumbleweed" ]]; then
-    echo "Your OS is OpenSUSE Tumbleweed"
+    LOGI "Your OS is OpenSUSE Tumbleweed"
 elif [[ "${release}" == "openEuler" ]]; then
     if [[ ${os_version} -lt 2203 ]]; then
-        echo -e "${red} Please use OpenEuler 22.03 or higher ${plain}\n" && exit 1
+        LOGE "Please use OpenEuler 22.03 or higher"
+        exit 1
     fi
 elif [[ "${release}" == "centos" ]]; then
     if [[ ${os_version} -lt 8 ]]; then
-        echo -e "${red} Please use CentOS 8 or higher ${plain}\n" && exit 1
+        LOGE "Please use CentOS 8 or higher"
+        exit 1
     fi
 elif [[ "${release}" == "ubuntu" ]]; then
     if [[ ${os_version} -lt 2004 ]]; then
-        echo -e "${red} Please use Ubuntu 20 or higher version!${plain}\n" && exit 1
+        LOGE "Please use Ubuntu 20 or higher version!"
+        exit 1
     fi
 elif [[ "${release}" == "fedora" ]]; then
     if [[ ${os_version} -lt 36 ]]; then
-        echo -e "${red} Please use Fedora 36 or higher version!${plain}\n" && exit 1
+        LOGE "Please use Fedora 36 or higher version!"
+        exit 1
     fi
 elif [[ "${release}" == "amzn" ]]; then
     if [[ ${os_version} != "2023" ]]; then
-        echo -e "${red} Please use Amazon Linux 2023!${plain}\n" && exit 1
+        LOGE "Please use Amazon Linux 2023!"
+        exit 1
     fi
 elif [[ "${release}" == "debian" ]]; then
     if [[ ${os_version} -lt 11 ]]; then
-        echo -e "${red} Please use Debian 11 or higher ${plain}\n" && exit 1
+        LOGE "Please use Debian 11 or higher"
+        exit 1
     fi
 elif [[ "${release}" == "almalinux" ]]; then
     if [[ ${os_version} -lt 80 ]]; then
-        echo -e "${red} Please use AlmaLinux 8.0 or higher ${plain}\n" && exit 1
+        LOGE "Please use AlmaLinux 8.0 or higher"
+        exit 1
     fi
 elif [[ "${release}" == "rocky" ]]; then
     if [[ ${os_version} -lt 8 ]]; then
-        echo -e "${red} Please use Rocky Linux 8 or higher ${plain}\n" && exit 1
+        LOGE "Please use Rocky Linux 8 or higher"
+        exit 1
     fi
 elif [[ "${release}" == "ol" ]]; then
     if [[ ${os_version} -lt 8 ]]; then
-        echo -e "${red} Please use Oracle Linux 8 or higher ${plain}\n" && exit 1
+        LOGE "Please use Oracle Linux 8 or higher"
+        exit 1
     fi
 else
-    echo -e "${red}Your operating system is not supported by this script.${plain}\n"
+    LOGE "Your operating system is not supported by this script."
     echo "Please ensure you are using one of the following supported operating systems:"
     echo "- Ubuntu 20.04+"
     echo "- Debian 11+"
@@ -113,22 +134,22 @@ fi
 install_base() {
     case "${release}" in
     ubuntu | debian | armbian)
-        apt-get update && apt-get install -y -q wget curl tar tzdata
+        apt-get update && apt-get install -y -q wget curl tar tzdata socat fail2ban
         ;;
     centos | almalinux | rocky | ol)
-        yum -y update && yum install -y -q wget curl tar tzdata
+        yum -y update && yum install -y -q wget curl tar tzdata socat fail2ban
         ;;
     fedora | amzn)
-        dnf -y update && dnf install -y -q wget curl tar tzdata
+        dnf -y update && dnf install -y -q wget curl tar tzdata socat fail2ban
         ;;
     arch | manjaro | parch)
-        pacman -Syu && pacman -Syu --noconfirm wget curl tar tzdata
+        pacman -Syu && pacman -Syu --noconfirm wget curl tar tzdata socat fail2ban
         ;;
     opensuse-tumbleweed)
-        zypper refresh && zypper -q install -y wget curl tar timezone
+        zypper refresh && zypper -q install -y wget curl tar timezone socat fail2ban
         ;;
     *)
-        apt-get update && apt install -y -q wget curl tar tzdata
+        apt-get update && apt install -y -q wget curl tar tzdata socat fail2ban
         ;;
     esac
 }
@@ -140,10 +161,16 @@ gen_random_string() {
 }
 
 config_after_install() {
-    local existing_username=$(/usr/local/x-sl/x-sl setting -show true | grep -Eo 'Owner username: .+' | awk '{print $2}')
-    local existing_password=$(/usr/local/x-sl/x-sl setting -show true | grep -Eo 'Owner password: .+' | awk '{print $2}')
-    local existing_webBasePath=$(/usr/local/x-sl/x-sl setting -show true | grep -Eo 'webBasePath: .+' | awk '{print $2}')
-    local existing_port=$(/usr/local/x-sl/x-sl setting -show true | grep -Eo 'port: .+' | awk '{print $2}')
+    local settings_output=$(/usr/local/x-sl/x-sl setting -show true)
+    if [[ $? -ne 0 ]]; then
+        LOGE "Failed to retrieve X-SL settings. Please check if X-SL is installed correctly."
+        exit 1
+    fi
+
+    local existing_username=$(echo "$settings_output" | grep -Eo 'Owner username: .+' | awk '{print $2}')
+    local existing_password=$(echo "$settings_output" | grep -Eo 'Owner password: .+' | awk '{print $2}')
+    local existing_webBasePath=$(echo "$settings_output" | grep -Eo 'webBasePath: .+' | awk '{print $2}')
+    local existing_port=$(echo "$settings_output" | grep -Eo 'port: .+' | awk '{print $2}')
     local server_ip=$(curl -s https://api.ipify.org)
 
     if [[ ${#existing_webBasePath} -lt 4 ]]; then
@@ -155,44 +182,44 @@ config_after_install() {
             read -p "Would you like to customize the Panel Port settings? (If not, a random port will be applied) [y/n]: " config_confirm
             if [[ "${config_confirm}" == "y" || "${config_confirm}" == "Y" ]]; then
                 read -p "Please set up the X-SL panel port: " config_port
-                echo -e "${yellow}Your X-SL Panel Port is: ${config_port}${plain}"
+                LOGI "Your X-SL Panel Port is: ${config_port}"
             else
                 local config_port=$(shuf -i 1024-62000 -n 1)
-                echo -e "${yellow}Generated random port: ${config_port}${plain}"
+                LOGI "Generated random port: ${config_port}"
             fi
 
             /usr/local/x-sl/x-sl setting -username "${config_username}" -password "${config_password}" -port "${config_port}" -webBasePath "${config_webBasePath}"
-            echo -e "This is a fresh installation, generating random login info for security concerns:"
+            LOGI "This is a fresh installation, generating random login info for security concerns:"
             echo -e "╭───────────────────────────────────────────────────────────────────────────────────────╮"
-            echo -e "${green}Username: ${config_username}${plain}"
-            echo -e "${green}Password: ${config_password}${plain}"
-            echo -e "${green}Port: ${config_port}${plain}"
-            echo -e "${green}WebBasePath: ${config_webBasePath}${plain}"
-            echo -e "${green}Access URL: http://${server_ip}:${config_port}/${config_webBasePath}${plain}"
+            LOGI "Username: ${config_username}"
+            LOGI "Password: ${config_password}"
+            LOGI "Port: ${config_port}"
+            LOGI "WebBasePath: ${config_webBasePath}"
+            LOGI "Access URL: http://${server_ip}:${config_port}/${config_webBasePath}"
             echo -e "╰───────────────────────────────────────────────────────────────────────────────────────╯"
-            echo -e "${yellow}If you forgot your login info, you can type 'x-sl settings' to check${plain}"
+            LOGI "If you forgot your login info, you can type 'x-sl settings' to check"
         else
             local config_webBasePath=$(gen_random_string 15)
-            echo -e "${yellow}WebBasePath is missing or too short. Generating a new one...${plain}"
+            LOGI "WebBasePath is missing or too short. Generating a new one..."
             /usr/local/x-sl/x-sl setting -webBasePath "${config_webBasePath}"
-            echo -e "${green}New WebBasePath: ${config_webBasePath}${plain}"
-            echo -e "${green}Your Access URL: http://${server_ip}:${existing_port}/${config_webBasePath}${plain}"
+            LOGI "New WebBasePath: ${config_webBasePath}"
+            LOGI "Your Access URL: http://${server_ip}:${existing_port}/${config_webBasePath}"
         fi
     else
         if [[ "$existing_username" == "admin" && "$existing_password" == "admin" ]]; then
             local config_username=$(gen_random_string 10)
             local config_password=$(gen_random_string 10)
 
-            echo -e "${yellow}Default credentials detected. Security update required...${plain}"
+            LOGI "Default credentials detected. Security update required..."
             /usr/local/x-sl/x-sl setting -username "${config_username}" -password "${config_password}"
-            echo -e "Generated new random login credentials:"
+            LOGI "Generated new random login credentials:"
             echo -e "╭┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈╮"
-            echo -e "${green}Username: ${config_username}${plain}"
-            echo -e "${green}Password: ${config_password}${plain}"
+            LOGI "Username: ${config_username}"
+            LOGI "Password: ${config_password}"
             echo -e "╰┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈╯"
-            echo -e "${yellow}If you forgot your login info, you can type 'x-sl settings' to check${plain}"
+            LOGI "If you forgot your login info, you can type 'x-sl settings' to check"
         else
-            echo -e "${green}Username, Password, and WebBasePath are properly set. Exiting...${plain}"
+            LOGI "Username, Password, and WebBasePath are properly set. Exiting..."
         fi
     fi
 
@@ -205,13 +232,13 @@ install_x-sl() {
     if [ $# == 0 ]; then
         tag_version=$(curl -Ls "https://api.github.com/repos/MasterHide/X-SL/releases/latest" | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
         if [[ ! -n "$tag_version" ]]; then
-            echo -e "${red}Failed to fetch X-SL version, it may be due to GitHub API restrictions, please try it later${plain}"
+            LOGE "Failed to fetch X-SL version. Please check your internet connection or GitHub API status."
             exit 1
         fi
-        echo -e "Got x-sl latest version: ${tag_version}, beginning the installation..."
+        LOGI "Got x-sl latest version: ${tag_version}, beginning the installation..."
         wget -N --no-check-certificate -O /usr/local/x-sl-linux-$(arch).tar.gz https://github.com/MasterHide/X-SL/releases/download/${tag_version}/x-sl-linux-$(arch).tar.gz
         if [[ $? -ne 0 ]]; then
-            echo -e "${red}Downloading x-sl failed, please be sure that your server can access GitHub ${plain}"
+            LOGE "Downloading x-sl failed, please be sure that your server can access GitHub"
             exit 1
         fi
     else
@@ -220,15 +247,19 @@ install_x-sl() {
         min_version="1.0.1"
 
         if [[ "$(printf '%s\n' "$min_version" "$tag_version_numeric" | sort -V | head -n1)" != "$min_version" ]]; then
-            echo -e "${red}Please use a newer version (at least v1.0.0). Exiting installation.${plain}"
+            LOGE "Please use a newer version (at least v1.0.1). Exiting installation."
             exit 1
         fi
 
         url="https://github.com/MasterHide/X-SL/releases/download/${tag_version}/x-sl-linux-$(arch).tar.gz"
-        echo -e "Beginning to install x-sl $1"
+        if ! curl --output /dev/null --silent --head --fail "$url"; then
+            LOGE "Invalid URL: $url. Please check if the version exists."
+            exit 1
+        fi
+        LOGI "Beginning to install x-sl $1"
         wget -N --no-check-certificate -O /usr/local/x-sl-linux-$(arch).tar.gz ${url}
         if [[ $? -ne 0 ]]; then
-            echo -e "${red}Download x-sl $1 failed, please check if the version exists ${plain}"
+            LOGE "Download x-sl $1 failed, please check if the version exists"
             exit 1
         fi
     fi
@@ -239,6 +270,10 @@ install_x-sl() {
     fi
 
     tar zxvf x-sl-linux-$(arch).tar.gz
+    if [[ $? -ne 0 ]]; then
+        LOGE "Failed to extract x-sl archive. Please check the file integrity."
+        exit 1
+    fi
     rm x-sl-linux-$(arch).tar.gz -f
     cd x-sl
     chmod +x x-sl
@@ -257,9 +292,24 @@ install_x-sl() {
     config_after_install
 
     systemctl daemon-reload
+    if [[ $? -ne 0 ]]; then
+        LOGE "Failed to reload systemd daemon. Please check the systemd configuration."
+        exit 1
+    fi
+
     systemctl enable x-sl
+    if [[ $? -ne 0 ]]; then
+        LOGE "Failed to enable x-sl service. Please check the systemd configuration."
+        exit 1
+    fi
+
     systemctl start x-sl
-    echo -e "${green}x-sl ${tag_version}${plain} installation completed,running now..."
+    if [[ $? -ne 0 ]]; then
+        LOGE "Failed to start x-sl service. Please check the logs for more details."
+        exit 1
+    fi
+
+    LOGI "x-sl ${tag_version} installation completed, running now..."
     echo -e ""
     echo -e "┌───────────────────────────────────────────────────────┐
 │  ${blue}x-sl control menu usages (subcommands):${plain}              │
@@ -281,6 +331,16 @@ install_x-sl() {
 └───────────────────────────────────────────────────────┘"
 }
 
-echo -e "${green}Running...${plain}"
+cleanup() {
+    rm -rf /usr/local/x-sl-linux-$(arch).tar.gz
+    rm -rf /usr/local/x-sl/
+    systemctl stop x-sl
+    systemctl disable x-sl
+    rm -f /etc/systemd/system/x-sl.service
+}
+
+trap cleanup EXIT
+
+LOGI "Running..."
 install_base
 install_x-sl $1
